+++ example/src/components/layouts/AuthLayout.tsx
@@ -3,6 +3,7 @@ import { Outlet, Navigate } from "react-router-dom";
 
 const AuthLayout = () => {
   const isAuthenticated = useIsAuthenticated();
+  if (isAuthenticated === null) return <span>Authenticating</span>;
   return isAuthenticated ? <Outlet /> : <Navigate to="/login" replace />;
 };